Implement wrapping with abbreviation with Tab
