+++ EmmetVS/Options/GeneralOptions.cs
@@ -10,7 +10,8 @@ public class GeneralOptions : BaseOptionModel<GeneralOptions>
 {
     private bool _enable = true;
     private bool _enableAdvanced = false;
-    private bool _enableTabKey = true;
+    private bool _enableExpandWithTabKey = true;
+    private bool _enableWrapWithTabKey = true;
 
     /// <summary>
     /// Gets or sets the property changed event handler.
@@ -64,15 +65,35 @@ public class GeneralOptions : BaseOptionModel<GeneralOptions>
     [DisplayName("Enable Expanding Abbreviation With Tab Key")]
     [Description("Select whether to enable expanding abbreviations with tab key.")]
     [DefaultValue(true)]
-    public bool EnableTabKey
+    public bool EnableExpandWithTabKey
     {
-        get => _enableTabKey;
+        get => _enableExpandWithTabKey;
         set
         {
-            if (_enableTabKey != value)
+            if (_enableExpandWithTabKey != value)
             {
-                _enableTabKey = value;
-                OnPropertyChanged(nameof(EnableTabKey));
+                _enableExpandWithTabKey = value;
+                OnPropertyChanged(nameof(EnableExpandWithTabKey));
+            }
+        }
+    }
+
+    /// <summary>
+    /// Gets or sets whether to enable wrapping with abbreviation with tab key.
+    /// </summary>
+    [Category("General Options")]
+    [DisplayName("Enable Wrapping With Abbreviation With Tab Key")]
+    [Description("Select whether to enable wrapping with abbreviation with tab key.")]
+    [DefaultValue(true)]
+    public bool EnableWrapWithTabKey
+    {
+        get => _enableWrapWithTabKey;
+        set
+        {
+            if (_enableWrapWithTabKey != value)
+            {
+                _enableWrapWithTabKey = value;
+                OnPropertyChanged(nameof(EnableWrapWithTabKey));
             }
         }
     }
